Pull the MapDomainSpec method out to a new package
The process of mapping domain specs is going to gain
significantly more functionality. It will also be used
by more than just virt-handler, so isolate it from
the rest of the virt-handler code in its own package.
